requirements.txt
Django>=4.0,<5.0
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
djangorestframework>=3.14.0  # If using API
gunicorn>=20.1.0  # For production